+++ glob/network.go
@@ -1,0 +1,28 @@
+package glob
+
+import (
+	"fmt"
+	"net"
+	"strings"
+)
+
+func GetMacAddrs() (macAddrs []string) {
+	netInterfaces, err := net.Interfaces()
+	if err != nil {
+		fmt.Printf("fail to get net interfaces: %v", err)
+		return macAddrs
+	}
+
+	for _, netInterface := range netInterfaces {
+		macAddr := netInterface.HardwareAddr.String()
+		if len(macAddr) == 0 {
+			continue
+		}
+		macAddrs = append(macAddrs, macAddr)
+	}
+	return macAddrs
+}
+
+func GetHashMacAddrs() string {
+	return HashMd5(strings.Join(GetMacAddrs(), ","))
+}